// src/data/model/TVShow/TVShowOverview.ts
import {media_type} from "../../Datasource/Config";

export default interface TVShowOverview {
    adult: boolean;
    backdrop_path: string;
    id: number;
    name: string;
    original_language: string;
    original_name: string;
    overview: string;
    poster_path: string;
    media_type: media_type;
    genre_ids: number[];
    popularity: number;
    first_air_date: string;
    vote_average: number;
    vote_count: number;
    origin_country: string[];
}